# artifacts/desk-openapi.yaml
openapi: 3.0.3
info:
  title: Macro Agent desk API
  version: 1.0.0
paths:
...
  /api/v1/models/:
    get:
      operationId: compilation_model_catalog
      description: Development-only explicit model catalogue from the configured provider.
        Requires local settings, the model-compilation gate and a development/test
        database. Prices are advertised metadata, not final charges. The API key is
        never returned. No model is selected automatically.
      tags:
      - internal compilation
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelCatalogResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
...
  /api/v1/theses/{thesis_id}/compilation-commands/{command_id}/:
    get:
      operationId: get_compilation_command
      description: Read-only recovery by the command UUID saved before transmission.
        No provider credentials, catalogue lookup, inference or POST fallback. Historical
        results are shown alongside current draft disposition.
      parameters:
      - in: path
        name: command_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: thesis_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - internal compilation
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompilationResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '415':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
  /api/v1/theses/{thesis_id}/compilations/{attempt_id}/:
    get:
      operationId: get_compilation
      description: Read an owner-scoped saved compilation attempt and its current
        draft disposition. A running or unknown attempt is visible without initiating
        inference or retries. The response does not assert factual verification, approval
        or continuous monitoring.
      parameters:
      - in: path
        name: attempt_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: thesis_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - internal compilation
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompilationResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
  /api/v1/theses/{thesis_id}/compile/:
    post:
      operationId: compile_thesis
      description: Development-only one-call text-grounded compilation using an explicit
        catalogue model. The authenticated owner supplies no credentials or provider
        configuration in this request. The result is a suggested draft interpretation
        with separate unverified questions and hypotheses. Exact thesis text and current
        approval remain unchanged. Explicit user approval is still required. There
        are no fact sources, current macro context, model tools or monitoring daemon.
        Inspect status even on HTTP 200; saved command retries do not initiate another
        model call.
      parameters:
      - in: path
        name: thesis_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - internal compilation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompileThesisRequestRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompilationResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '415':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
...
  /api/v1/theses/{thesis_id}/refinement-commands/{command_id}/:
    get:
      operationId: get_thesis_refinement_command
      description: Read-only answer-submission recovery by saved command UUID. No
        catalogue, credentials, model request or POST fallback.
      parameters:
      - in: path
        name: command_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: thesis_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - internal compilation
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefinementResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '415':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
  /api/v1/theses/{thesis_id}/refinements/:
    post:
      operationId: save_thesis_refinement
      description: Save exact answers against questions from a current retained proposal,
        without a provider request, paid-call admission or thesis approval. The service
        derives question wording and bounded cumulative history. A new explicit compilation
        may reference this submission later.
      parameters:
      - in: path
        name: thesis_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - internal compilation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveRefinementRequestRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefinementResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '415':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
  /api/v1/theses/{thesis_id}/refinements/{refinement_id}/:
    get:
      operationId: get_thesis_refinement
      description: Read an owner-scoped saved answer submission without provider work
        or activation.
      parameters:
      - in: path
        name: refinement_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: thesis_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - internal compilation
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefinementResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '415':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
components:
  schemas:
...
    Approved:
      type: object
      properties:
        text_version:
          $ref: '#/components/schemas/TextVersion'
        interpretation:
          $ref: '#/components/schemas/Interpretation'
        approval:
          $ref: '#/components/schemas/Approval'
      required:
      - approval
      - interpretation
      - text_version
...
    CatalogModel:
      type: object
      properties:
        id:
          type: string
          maxLength: 255
        name:
          type: string
          maxLength: 512
        context_length:
          type: integer
          minimum: 1
          nullable: true
        input_price_usd_per_million:
          type: string
          nullable: true
          description: Advertised USD per million input tokens, or unknown. Not a
            final charge.
        output_price_usd_per_million:
          type: string
          nullable: true
          description: Advertised USD per million output tokens, or unknown. Not a
            final charge.
        capabilities:
          type: object
          additionalProperties:
            type: boolean
      required:
      - capabilities
      - context_length
      - id
      - input_price_usd_per_million
      - name
      - output_price_usd_per_million
...
    Compilation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          allOf:
          - $ref: '#/components/schemas/CompilationStatusEnum'
          description: |-
            Explicit attempt disposition. HTTP 200 does not mean compilation succeeded. An unknown outcome may have incurred provider cost and must not be retried implicitly.

            * `running` - running
            * `outcome_unknown` - outcome_unknown
            * `compiled` - compiled
            * `stale` - stale
            * `failed` - failed
        provider:
          $ref: '#/components/schemas/ModelProviderEnum'
        model_id:
          type: string
          maxLength: 255
        input_text_version:
          $ref: '#/components/schemas/CompilationTextInput'
        schema_version:
          type: string
        prompt_version:
          type: string
        refinement_id:
          type: string
          format: uuid
          nullable: true
        parent_attempt_id:
          type: string
          format: uuid
          nullable: true
        refinement_inputs:
          type: array
          items:
            $ref: '#/components/schemas/RefinementInput'
        created_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
          nullable: true
        document:
          allOf:
          - $ref: '#/components/schemas/CompilationDocument'
          nullable: true
        interpretation_version_id:
          type: string
          format: uuid
          nullable: true
        is_current_draft:
          type: boolean
          description: Protected response snapshot of whether this interpretation
            remains the latest draft. A draft is not user approval or active monitoring.
        usage:
          $ref: '#/components/schemas/CompilationUsage'
        reported_cost_usd:
          type: string
          nullable: true
        estimated_cost_usd:
          type: string
          nullable: true
          description: Estimate from recorded catalogue pricing and reported token
            usage; unknown is null.
        reported_model:
          type: string
          nullable: true
        provider_request_id:
          type: string
          nullable: true
        latency_ms:
          type: integer
          minimum: 0
          nullable: true
        stop_reason:
          type: string
          nullable: true
        limitations:
          type: array
          items:
            type: string
      required:
      - created_at
      - document
      - estimated_cost_usd
      - finished_at
      - id
      - input_text_version
      - interpretation_version_id
      - is_current_draft
      - latency_ms
      - limitations
      - model_id
      - parent_attempt_id
      - prompt_version
      - provider
      - provider_request_id
      - refinement_id
      - refinement_inputs
      - reported_cost_usd
      - reported_model
      - schema_version
      - status
      - stop_reason
      - usage
    CompilationDocument:
      type: object
      properties:
        interpretation:
          $ref: '#/components/schemas/InterpretationInput'
        grounding:
          type: array
          items:
            $ref: '#/components/schemas/CompilationGrounding'
        refinement_issues:
          type: array
          items:
            $ref: '#/components/schemas/CompilationRefinementIssue'
        agent_hypotheses:
          type: array
          items:
            $ref: '#/components/schemas/CompilationHypothesis'
        counter_case:
          type: string
          nullable: true
          maxLength: 2000
        review_card:
          allOf:
          - $ref: '#/components/schemas/ReviewSections'
          description: Required for current v2 model output; absent in retained historical
            v1 documents. The complete card combines these sections with interpretation,
            counter-case and questions.
      required:
      - agent_hypotheses
      - counter_case
      - grounding
      - interpretation
      - refinement_issues
      additionalProperties: false
    CompilationGrounding:
      type: object
      properties:
        field:
          $ref: '#/components/schemas/CompilationGroundingFieldEnum'
        index:
          type: integer
          maximum: 31
          minimum: 0
          nullable: true
        input_id:
          type: string
          description: Named original thesis or exact trader-answer input. Absent
            only in historical v1 documents.
          maxLength: 100
        exact_quote:
          type: string
          maxLength: 1000
      required:
      - exact_quote
      - field
      - index
      additionalProperties: false
    CompilationGroundingFieldEnum:
      enum:
      - drivers
      - horizon
      - invalidation_signposts
      type: string
      description: |-
        * `drivers` - drivers
        * `horizon` - horizon
        * `invalidation_signposts` - invalidation_signposts
...
    CompilationResponse:
      type: object
      properties:
        compilation:
          $ref: '#/components/schemas/Compilation'
        thesis:
          $ref: '#/components/schemas/ThesisDetail'
      required:
      - compilation
      - thesis
    CompilationStatusEnum:
      enum:
      - running
      - outcome_unknown
      - compiled
      - stale
      - failed
      type: string
      description: |-
        * `running` - running
        * `outcome_unknown` - outcome_unknown
        * `compiled` - compiled
        * `stale` - stale
        * `failed` - failed
    CompilationTextInput:
      type: object
      properties:
        id:
          type: string
          format: uuid
        parent_version_id:
          type: string
          format: uuid
          nullable: true
        exact_text:
          type: string
        text_digest:
          type: string
        created_at:
          type: string
          format: date-time
      required:
      - created_at
      - exact_text
      - id
      - parent_version_id
      - text_digest
    CompilationUsage:
      type: object
      properties:
        prompt_tokens:
          type: integer
          minimum: 0
          nullable: true
        completion_tokens:
          type: integer
          minimum: 0
          nullable: true
        total_tokens:
          type: integer
          minimum: 0
          nullable: true
      required:
      - completion_tokens
      - prompt_tokens
      - total_tokens
    CompileThesisRequestRequest:
      type: object
      properties:
        command_id:
          type: string
          format: uuid
          minLength: 1
          maxLength: 36
        refinement_id:
          type: string
          format: uuid
          nullable: true
          minLength: 1
          maxLength: 36
        expected_revision:
          type: integer
          maximum: 9223372036854775807
          minimum: 1
          format: int64
        provider_id:
          allOf:
          - $ref: '#/components/schemas/ModelProviderEnum'
          description: |-
            Provider identity shown by the catalogue, pinned against backend configuration.

            * `nanogpt` - nanogpt
            * `openrouter` - openrouter
            * `cheaperinference` - cheaperinference
        model_id:
          type: string
          minLength: 1
          description: Explicit model ID selected from the configured provider catalogue.
            No automatic routing.
          maxLength: 255
      required:
      - command_id
      - expected_revision
      - model_id
      - provider_id
      additionalProperties: false
...
    Draft:
      type: object
      properties:
        text_version:
          $ref: '#/components/schemas/TextVersion'
        interpretation:
          $ref: '#/components/schemas/Interpretation'
      required:
      - interpretation
      - text_version
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
        errors: {}
      required:
      - detail
...
    InterpretationInput:
      type: object
      properties:
        drivers:
          type: array
          items:
            type: string
            maxLength: 1000
          maxItems: 32
        horizon:
          type: string
          nullable: true
          maxLength: 1000
        invalidation_signposts:
          type: array
          items:
            type: string
            maxLength: 1000
          maxItems: 32
      required:
      - drivers
      - horizon
      - invalidation_signposts
      additionalProperties: false
...
    ModelCatalogResponse:
      type: object
      properties:
        provider:
          $ref: '#/components/schemas/ModelProviderEnum'
        compilation_enabled:
          type: boolean
        credentials_configured:
          type: boolean
        fetched_at:
          type: string
          format: date-time
        models:
          type: array
          items:
            $ref: '#/components/schemas/CatalogModel'
      required:
      - compilation_enabled
      - credentials_configured
      - fetched_at
      - models
      - provider
...
    ModelProviderEnum:
      enum:
      - nanogpt
      - openrouter
      - cheaperinference
      type: string
      description: |-
        * `nanogpt` - nanogpt
        * `openrouter` - openrouter
        * `cheaperinference` - cheaperinference
    MonitoringEnum:
      enum:
      - not_configured
      type: string
      description: '* `not_configured` - not_configured'
...
    Refinement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        command_id:
          type: string
          format: uuid
        parent_attempt_id:
          type: string
          format: uuid
        text_version_id:
          type: string
          format: uuid
        input_text_version:
          $ref: '#/components/schemas/CompilationTextInput'
        expected_revision:
          type: integer
        answers:
          type: array
          items:
            $ref: '#/components/schemas/RefinementAnswer'
        cumulative_inputs:
          type: array
          items:
            $ref: '#/components/schemas/RefinementInput'
        created_at:
          type: string
          format: date-time
        replayed:
          type: boolean
        is_current_context:
          type: boolean
          description: Protected snapshot of whether this saved input can be used
            for the current draft. Saving answers does not approve or amend the thesis.
      required:
      - answers
      - command_id
      - created_at
      - cumulative_inputs
      - expected_revision
      - id
      - input_text_version
      - is_current_context
      - parent_attempt_id
      - replayed
      - text_version_id
    RefinementAnswer:
      type: object
      properties:
        question_index:
          type: integer
          maximum: 31
          minimum: 0
        exact_answer:
          type: string
          maxLength: 2000
      required:
      - exact_answer
      - question_index
      additionalProperties: false
    RefinementAnswerRequest:
      type: object
      properties:
        question_index:
          type: integer
          maximum: 31
          minimum: 0
        exact_answer:
          type: string
          minLength: 1
          maxLength: 2000
      required:
      - exact_answer
      - question_index
      additionalProperties: false
    RefinementInput:
      type: object
      properties:
        question_index:
          type: integer
          maximum: 31
          minimum: 0
        exact_answer:
          type: string
          maxLength: 2000
        input_id:
          type: string
          maxLength: 100
        parent_attempt_id:
          type: string
          format: uuid
          maxLength: 36
        question:
          type: string
          maxLength: 1000
      required:
      - exact_answer
      - input_id
      - parent_attempt_id
      - question
      - question_index
      additionalProperties: false
    RefinementResponse:
      type: object
      properties:
        refinement:
          $ref: '#/components/schemas/Refinement'
        thesis:
          $ref: '#/components/schemas/ThesisDetail'
      required:
      - refinement
      - thesis
...
    SaveRefinementRequestRequest:
      type: object
      properties:
        command_id:
          type: string
          format: uuid
          minLength: 1
          maxLength: 36
        expected_revision:
          type: integer
          maximum: 9223372036854775807
          minimum: 1
          format: int64
        parent_attempt_id:
          type: string
          format: uuid
          minLength: 1
          maxLength: 36
        answers:
          type: array
          items:
            $ref: '#/components/schemas/RefinementAnswerRequest'
      required:
      - answers
      - command_id
      - expected_revision
      - parent_attempt_id
      additionalProperties: false
...
    ThesisDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        revision:
          type: integer
        created_at:
          type: string
          format: date-time
        changed_at:
          type: string
          format: date-time
        draft:
          $ref: '#/components/schemas/Draft'
        approved:
          allOf:
          - $ref: '#/components/schemas/Approved'
          nullable: true
        monitoring:
          $ref: '#/components/schemas/MonitoringEnum'
      required:
      - approved
      - changed_at
      - created_at
      - draft
      - id
      - monitoring
      - revision
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid